+++ apps-client/tarefas/free/dialogs/edit-task.js
@@ -42,6 +42,11 @@ app.routes.dialog('/editar-tarefa/:id', function (params, data) {
         reminderOptions = [],
 
         /**
+         * Lista de ui.option de usuários
+         */
+        userOptions = [],
+
+        /**
          * Fieldset
          */
         fieldset,
@@ -63,6 +68,17 @@ app.routes.dialog('/editar-tarefa/:id', function (params, data) {
                     value : categories[i]._id,
                     label : categories[i].color || 'blue',
                     clicked : categories[i]._id === task.category
+                }));
+            }
+        }
+
+        /* Input com os usuários */
+        for (var i in app.config.users) {
+            if (app.config.users.hasOwnProperty(i)) {
+                userOptions.push(new app.ui.inputOption({
+                    legend  : app.config.users[i].name,
+                    value   : app.config.users[i]._id,
+                    clicked : task.user === app.config.users[i]._id
                 }));
             }
         }
@@ -155,6 +171,15 @@ app.routes.dialog('/editar-tarefa/:id', function (params, data) {
             value : task.description
         });
 
+        /* responsável */
+        fields.user = new app.ui.inputSelector({
+            name : 'user',
+            type : 'single',
+            legend  : 'Responsável',
+            options : userOptions,
+            filterable : true
+        });
+
         /* fieldset */
         fieldset = new app.ui.fieldset({
             legend : 'Tarefa'
@@ -171,6 +196,7 @@ app.routes.dialog('/editar-tarefa/:id', function (params, data) {
         fieldset.fields.add(fields.reminder);
         fieldset.fields.add(fields.recurrence);
         fieldset.fields.add(fields.description);
+        fieldset.fields.add(fields.user);
         app.ui.form.fieldsets.add(fieldset);
 
         fields.title.focus();
@@ -183,7 +209,8 @@ app.routes.dialog('/editar-tarefa/:id', function (params, data) {
                 category : fields.category.value()[0],
                 important : fields.important.value()[0] === 'important',
                 recurrence : fields.recurrence.value()[0],
-                description : fields.description.value()
+                description : fields.description.value(),
+                user : fields.user.value()[0]
             };
 
             if (task.done) {