+++ apps-client/ee/free/changelog.md
@@ -1,7 +1,12 @@
 Changelog
 =========
 
-## 2.1
+## 2.2
+
+Features
+- Filtragem de usuário no painel de controle
+
+## 2.1 (2013-04-12)
 
 Features
 - Alterar senha de usuários